+++ src/form/BasicButton/BasicButtonTypeMap.ts
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { BoxTypeMap } from '@layout/Box';
+
+type BasicButtonTypeMap<P = {}, D extends React.ElementType = 'button'> = BoxTypeMap<P & {
+  variant?: 'text' | 'contain' | 'outline';
+  color?: 'primary' | 'secondary' | 'default';
+  size?: 'xs' | 'sm' | 'md' | 'lg' | 'xl';
+  disabled?: boolean;
+}, D>
+
+export default BasicButtonTypeMap;
