+++ src/form/BasicButton/index.ts
@@ -1,0 +1,5 @@
+import BasicButton from './BasicButton';
+
+export type { default as BasicButtonProps } from './BasicButtonProps';
+export type { default as BasicButtonTypeMap } from './BasicButtonTypeMap';
+export default BasicButton;
