+++ src/form/Button/index.ts
@@ -1,4 +1,5 @@
 import Button from './Button';
 
 export type { default as ButtonProps } from './ButtonProps';
+export type { default as ButtonTypeMap } from './ButtonTypeMap';
 export default Button;
