+++ src/form/CircleButton/index.ts
@@ -1,0 +1,5 @@
+import CircleButton from './CircleButton';
+
+export type { default as CircleButtonProps } from './CircleButtonProps';
+export type { default as CircleButtonTypeMap } from './CircleButtonTypeMap';
+export default CircleButton;
