+++ src/form/index.ts
@@ -1,2 +1,5 @@
 export { default as Button } from './Button';
-export type { ButtonProps } from './Button';
+export type { ButtonProps, ButtonTypeMap } from './Button';
+
+export { default as IconButton } from './CircleButton';
+export type { CircleButtonProps, CircleButtonTypeMap } from './CircleButton';
