+++ tailwind.config.js
@@ -215,6 +215,7 @@ module.exports = {
       xl: '0.75rem',
       '2xl': '1rem',
       '3xl': '1.5rem',
+      circle: '50%',
       full: '9999px',
     },
     borderSpacing: ({ theme }) => ({
